llm/a2a_adk/google-a2a-simpleAgent/pyproject.toml
[project]
name = "google-a2a-simpleagent"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "a2a-sdk>=0.2.4",
    "uvicorn>=0.34.2",
]
